Incomplete comparison skeleton - Luna
A skeleton code off of the primary observatory computer - Incomplete but probably still useful for future reference. I think there's a similar bit of code on my laptop, which is why I note what computer I found it on. The version/code on the laptop is likely the newer version but is also incomplete as far as I know